+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Terminate the process when one of the services crashes. ([#4249](https://github.com/getsentry/relay/pull/4249))
 - Don't propagate trace sampling decisions from SDKs ([#4265](https://github.com/getsentry/relay/pull/4265))
+- Rate limit profile chunks. ([#4270](https://github.com/getsentry/relay/pull/4270))
 
 **Features**: